+++ apps/merchant-mobile/src/assets/icons/index.ts
@@ -51,5 +51,6 @@ export { InfoIcon } from "./icons/info.icon";
 export { MoreHorizontalIcon } from "./icons/more-horizontal.icon";
 export { MoreVerticalIcon } from "./icons/more-vertical.icon";
 export { PencilIcon } from "./icons/pencil.icon";
+export { PlusIcon } from "./icons/plus.icon";
 export { UsersIcon } from "./icons/users.icon";
 export { ImageIcon } from "./icons/image.icon";